license: relicense "libnm-core/nm-dbus-interface.h" under LGPL-2.1+
We want to consistently use LGPL. Also, this file is part of libnm, which
requires its sources to be LGPL.

Also, this header file is supposed to be usable without libnm (e.g. for QT).

 $ git shortlog -n -s e055bdbbc3e7bb3ff3bc73860c4479f15bbfe7b2 -- libnm-core/nm-dbus-interface.h libnm-util/NetworkManager.h include/NetworkManager.h NetworkManager.h
    108    Dan Williams
     45    Thomas Haller
     34    Beniamino Galvani
     25    Dan Winship
     24    Lubomir Rintel
      9    Tambet Ingo
      7    Jiří Klimeš
      5    Robert Love
      4    Daniel Gnoutcheff
      3    Pantelis Koukousoulas
      2    Benjamin Berg
      2    Marius Vollmer
      1    Alfonso Sánchez-Beato
      1    Christopher Aillon
      1    Daniel Drake
      1    David Bauer
      1    Giovanni Campagna
      1    Jan Luebbe
      1    Javier Arteaga
      1    Jiri Pirko
      1    Marc Ruiz
      1    Rodrigo Moya
      1    Thomas Bechtold
      1    Thomas Graf
      1    Timothy Redaelli
      1    luz.paz

This is a large number of contributors. Changing the license here is
problematic. However, there is also the stance that simple header files
are not copyrightable. Maybe more importantly, the vast majority of
contributions come from people who would agree with this change and/or
were employed by Red Hat at the time (possibly the copyright lies with
Red Hat).

See contributions done by not the usual suspects:

  $ git log --stat -p --no-walk $(git log --pretty='%H:%an <%ae>' e055bdbbc3e7bb3ff3bc73860c4479f15bbfe7b2 -- libnm-core/nm-dbus-interface.h libnm-util/NetworkManager.h include/NetworkManager.h NetworkManager.h  | grep '\(dcbw\|bgalvani\|jklimes\|thaller\)@redhat.com\|danw@gnome.org\|lkundrak@v3.sk' -v | sed 's/:.*//') -- libnm-core/nm-dbus-interface.h libnm-util/NetworkManager.h include/NetworkManager.h NetworkManager.h