nm-policy: add new pending action when connection fails to prevent wrong startup complete
Add a pending action "waiting-for-autoactivation" when the device fails in activating
a connection, this action will later be removed when NM will try to activate a new
connection for the device or if the device can no longer be autoactivated.

This is needed as during the 'failed'->'disconnected' the pending action 'activation-*'
for the device is removed resulting in an empty pending_actions list which then
triggers 'check_if_startup_complete()' that will find no pending action and mark
startup as complete even if the device could have been activated with another connection.

https://bugzilla.redhat.com/show_bug.cgi?id=1759956